More work on an annotative encoder: picked out static fields, thoughts on where and how to define the "class" field to be used in encoding (which, incidentally, might be a really convenient answer to some questions about hintable codecs that haven't really gone away), test cases for actually selecting different fields with different args to the encoder that look up the values in the annotation...
For some bizzare reason all the constructors on the UnencodableException class were private; fixed.

Added a contains(String[] $arr, String $query) method to the Arr class.  It uses the equals(*) method... this is somewhat irksome to me, but apparently strings aren't always interned the way I thought they were on compile time.  (Of course, I should have a separate method for dealing with interned strings that is named appropriately anyway, but I was still a little shocked when my first try that assumed interning didn't work).