Annotative encoder is now paying some attention to type, though there's more to do there.
I'm going to go back to Eon stuff and see if I can't actually remove some genericness (for once... sounds odd coming out of my mouth).  I don't think the generic types on that thing really do much more than give subclasses a not-so-gentle reminder to override methods that are supposed to return their own type instead of general Eon* types, and that's definitely not worth the mass of added verbosity.