Okay, I have to be honest -- I've broken everything.
EonCodec needs to convey some special information to encoders that utilize it (namely so that the encoders don't have to specialize to specific encodings; or, in other words, so that those factories you handed to the EonCodec constructor actually -do- something).  Problem: the Encoder and Decoder types only give you the type information Codec<$CODE>.  Solution 1: add a ton of new methods to the Codec interface.  Not a viable solution: Codec doesn't want to be that complicated, and outright can't be that complicated in this case (it has no concept of objects vs arrays; that's domain knowledge to Eon stuff that's not necessarily portable).  Solution 2: do dirty typecasts in all of the implementers of Encoder and Decoder to transform Codec<EonObject> into EonCodec.  Legal, but extremely ishy: involves an extra line of boiler plate codec nearly 100% of the time, and results in systems that throw ClassCastException without any warning built in by the type system.  Solution 3: Make a -third- generic type argument for the Encoder and Decoder interfaces.  This involves changing -everything-, including vast swaths of Codec and CodecImpl, and as such scares me shitless.  Also, it just annoys me; it should be trivial to imply the code type from the codec type... except for the part where it can't actually be done in java.