All annotative codec stuff is moved out of experimental and into trunk (I'm contracting their names from ReflectiveAnnotative[En/De]coder to RAE and RAD because, well, it's sounds rad).
Old tests are currently failing to decode because of access problems on constructors.  I'm going to replace the old scheme of just praying for a default constructor with a scheme that requires a constructor with args that you should never see in the wild, thus avoiding inadvertant collisions as much as possible.