Added a simple test of the chill functions in util.X just as a sanity thing because of the BS that mcon has been giving me lately (there were no problems here).
Revised some documentation on ReadHead for greater precision.

Created WorkQueue, a class that combines a ConcurrentLinkedQueue with a Semaphore and some other locks in order to provide blocking reads and atomicity of some other operations that have previously been inconvenient and adhoc to guarantee.  I expect to use it both in mcon's primary node pusher as well as in the Pipe implementation of ReadHead (and possibily as the standard buffer in all ReadHead implementations, for that matter).