Made DomContactJso's condense(String...) method actually work correctly for line breaks.
Rearranged the hierarchy of top level objects a bit so that you can actually set attributes on the real body now.

Moved the World class to ahs.applet package.  It's still an extremely dubious abstraction -- there ought to be a diverse interface for logging and such that it provides -- but so be it.

World sets body attributes so it's... sane, makes the applet itself hidden, and sets up the main div to be big.

I'm still toying with ideas about making the entire range of common attributes and their valid keys accessible through a giant load of enums and having DomContact implementers translate them into the real strings, but that's not something I want to put a huge amount of effort into at the moment, and I'll stay supporting the current string based system for flexibility and backwards compatability pretty much forever anyway.