Well, it was a long journey, and I went places I never intended or wanted to go, but the old tests for byte stream read head are passing again.  I ended up changing some things that returned $T[] to List<$T> instead because I was having some truly surreal type casting errors -- a method that was just a one-line delegate to another method with the exact same interface (literally, they implemented the same interface for the same type) was throwing ClassCastException.
Added the obviously needed isClosed and close methods to WriteHead, but I'm still kind of scared of myself with that one there.  A huge part of me things that writers should just never buffer; it's not their job... it's either on the reader's head or in some cases on the head of the OS kernel itself.  Another part of me is a little scared of making such a sweeping declaration for all time, and a huge chunk of me is still weirded out by how the exception flow is completely unlike ReadHead.  Anyway, I think I just officially decided that I'm not providing a flush method.  ReadHead and WriteHead are supposed to operate on semantically significant objects; that's a their entire point.  Thus it seems only reasonable that a flush is implicit 100% of the time when writing to them for any reason.