Cleared up the delicacies of closing and exceptions resulting in closing in ReadHeadStreamToByteBuffer.
I was about to try to make an adapter that would let me use 90% of the same code for the ByteBuffer thing as well as the String thing and probably the EonObject thing.  It's a little scary, but aside from some bits of the pump code and the close method, nothing i'm currently dealing with actually needs to know a thing about the "base".  I doubt all of it will hold if I try to use the same thing on nonblocking nio Channel stuff, but I guess we'll give it a go and see how far it can get.

Now I want cake.