Made a ReadHead that's effectively a selectable server socket.
I realized the WriteHead on Pipe made some invalid assumptions about the internal queue's addAll functionality, which I fixed.

I'm currently... kinda frustrated with the crappy design of java's NIO.  (Not that that's new.  I've been working on this read/writehead stuff for weeks for a reason, after all.)  Why is it so hard (read: impossible) to have blocking writes and nonblocking reads?  That seems totally reasonable to me.  There's no fucking reason I should have to expose myself to working with the TCP stack's write buffers just because I want to be able to do a nonblocking read.  Jesus.