Added a "Flow" class that pairs a ReadHead and a Writehead.
Added a Translator implementation that goes from ByteBuffer to EonObject via an EonCodec.

Made EonReadHead a trivial extention of the general ReadHeadTranslator.

But I'm still flustered.  I need to put a serious rearrangement in the buffer strategies of ReadHead stuff, because it really should be going on only at the top level if threading is to be consistent through levels of a stack, and I don't think I'm going to have time for that scale of rethink for a while.