Moved the Translator for ByteBuffer to Eon into the actual Eon class to reduce clutter, as well as made one going the other direction.
Made a new method of Arr that will take any ByteBuffer and spawn a new one that can be read directly as an array.  It's kind of an odd thing to do, but it comes up enough when I'm cringing about efficiency that I figured it's probably handy to have ready.