Shuffled the names of the ReadHead adapters around (I'm not entirely willing to just toss the old one out at this point, despite the fact I certain don't intend to continue building on it).
Made Pipe throw an IllegalStateException when attempting to write to a pipe that has been closed.  (Considered ChannelClosedException among other things, but I find the import of java.nio.channels.* to be ishy, and I also don't really think there's any good reason for this to be anything other than a runtime exception, since in the case of internal pipes you really ought to know better.)

Rearranged the innards of ReadHeadAdapter (what used to be ''StackAdapter) severely.  That same adapter will now support both blocking and nonblocking subclasses via a slew of nested (and recursively) classes.