Made a massively more thorough set of test cases for eon codecs (and unlike previous codec tests, this one is actually applicable to anything that conforms to eon, instead of just specific implementations).  This was mostly spurred by the realization that my previous test cases for this area weren't very consistent in covering both encoding and serialization in all situations, and the distinction can matter.
Fixed the codec in EbonArray to use bytes instead of chars where applicable -- I guess I had forgotten to do this when I did EbonObjects, and it was actually causing a full break of the system since the array would put a two-byte switch on the front of itself when encoding, and then the object would try to pop a one-byte switch off when decoding and get super confused.  Anyway, it's all happily cleared up now and Ebon is passing the new codec tests completely again.