So... I spent a relatively sizable amount of time today deciding that the Bouncy Castle crypto library isn't anywhere near as perfect as I always dreamed it would be.  I already found one fairly major bug in their CBC block cipher's handling of IVs, and I'm -really- not a fan of the massive amount of defensive copies they do with byte arrays.  I mean, first of all, if a programmer isn't smart enough to understand where pointers to their byte arrays are going, they probably shouldn't be allowed near crypto projects anyway, and second of all, some of these copies are like copies-upon-copies-upon-copies where there's two or three purely inside the BC engine code even after a first one where data comes in from the rest of the application, and that's just silly.  AND... the way they set up argument passing between their block layer and their actual crypto engine, you can't change the IV on a cipher without also changing the key... which really makes one wonder why the hell they bother having a reset operation on their ciphers, since you can never actually sanely reuse them in their reset state anyway.
I'm debating how much of this is things that I want to attack/fix by a lot of fairly forcefull extending and overriding.  I think I'm definitely going to force AESEngine to check to make sure it's not already initialized with the key it's just been told to initialize to before it wastes time tearing the working key down and then building the exact same one again... and I think that in turn almost demands that I make some options available to myself for effectively sneaking data into something that's instanceof KeyParamter so that the AESEngine can make a super-cheap pointer-equality check on the key array instead of doing a larger equality check.

Some of this comes down to splitting hairs a little bit, but if I'm going to use this in MCON it needs to be direly efficient, and that inability to preserve cipher initialization just bugs the crap out of me.