Completed nearly all refactoring that was started in commit 0b8b1e02.
I widthdrew from the idea of separating out things into modules in the source -- instead I'll simply be separating buildable modules consistently by package (which also lines up with documentation generation better).

Most of the imports are corrected, but not quite all, and there are also most likely many extraneous imports that are now leftovers which will need to be removed.

Tests are all still todo.