Restructured Pipe to use a ReentrantLock internally instead of recycling the monitor of one of the existing objects for the write lock.
I did this because I thought I would need it for a continuing piece of work in the nonblocking IO APIs, but I'm not actually sure.  None the less, I'm keeping it because there doesn't seem to be any discernable performance difference and the increase in objects allocated is utterly insignificant.  It's not impossible that I'll restore the old implementation at a later date, though (they're completely interchangable).