PumperSelector in its internal event queues now handle write interest ops better, and include a new "attache" abstraction that connects several different Pumps to the same SelectionKey for handling the different types of operation.
This interface isn't completely final, unfortunately -- there is a severe inconvenience revolving around the fact that there's no sensible way to apply an interface that specificies WritableByteChannel (or Readable) at the same time as SelectableChannel (and Translator and Pump components only careabout the former, while Selector only cares about the latter).  This is why the older iterations of the PumperSelector interface typically allowed deregistrations to work based on Pump pointers, but that unfortunately implies the tradeoff of either doing linear walks of the selection key list (unacceptable) or keeping extra hashmaps around (unpleasant).