Building a WorkScheduler implementation with freely flexible priorities out of guts from Doug Lea's ScheduledThreadPoolExecutor.
It's not at all there yet; this is just DL's stuff trimmed down and stripped of ridiculous dependancies on package-protected things that make it impossible to just extend his work sanely.

I'm not entirely sure where this is going to end up before I call it done, but I have a very strong suspicion that I'm going to end up removing all methods having to do with recurrent or delayed scheduling from WorkScheduler that were mimicries of ScheduledThreadPoolExecutor, and instead replace the entire thing with a more elegant and consistent series of readiness reporting options on the WorkTarget interface.  Or something like that.