Admitting to the FuturePipe idea that's been sitting around on my local space forever.
I just noticed that it actually might be extremely easy to implement as a wrapper around an existing similar class (ExecutorCompletionService).  Despite the similarity, I'm still sufficiently attracted to the idea of making it look like a standard Flow that it's still well worth including such a class in the package.