Dicking around with terminal codes.  Color stuff is easy enough, of course -- I've done that before.
What I'm needing to take a little more seriously is stuff about positioning the cursor and detecting screen size and all that.

Also, I very much want to introduce some statefulness into this... for example, I want to be able to call a saveState() method on a Terminal and have it return to me a tuple of my color pallete, cursor position, and so on.

Oh, and at some point I should make a serious point of actually defining boxes, as well.  I shouldn't always have to do absolute coordinates (and also shouldn't always be allowed to).