It would appear that (with incredible rarity, and also with heisenbug principles in full force (adding logging drastically reduces the rate of occurance)) work is sometimes ending up in the unready pile instead of being updated to doneness and being properly finished and dropped.
I don't know how this is possible; I'd think that the doneness checks after the scheduler_power method finishes would be enough, since both before and after that check it's possible for concurrent updates to deal with finishing as well.  But there you are.  (And it's taking I-don't-even-know-how-many tens of thousands of runs -- which translates to hours of run time -- to catch one fail.)