Wrote up a quick little test of background and foreground coloring with the Palette system.
(I decided to put this in the main tree of the source instead of the test tree because I can see folks actually wanting to run this to see if their system works, even if they're not even remotely development inclined.  I'm also starting a habit from now on of putting classes with entry points (i.e. main methods) in "exe" packages.)