Shifted the weakly consistent logging for which tasks are actively running into locked sections.  The performance degradation of this move is undetectable above noise in my back-of-the-envelop testing, and it gives radically more information than before.
Most interesting of this new information is that I'm seeing a really large amount of time being spent with threads still holding on to tasks that are shifted into WAITING status already.  So... I'm surprised by that.  I think it must mean that we're spending way more time waiting to acquire that lock on line 142 than I anticipated.