Implemented the sifting on updates that makes the priorities in WorkSchedulerFlexiblePriority properly, well, flexible.
There were some serious suprises along the way that took me a while to realize the reasoning for.  Fixes exist, and they're not things anyone outside the library will ever see, but I'm entertaining the notion that these were a serious enough surprise to me that they might be worth refactoring for someday down the road.