ReadHeadSocketChannel is now mostly implemented as a WorkTarget + Translator based system.
There are a couple todo's in there, and I also definitely need to actually expose the WorkFuture, and so on.

This didn't actually turn out to use a TranslatingWorkTarget internally because of some bits with socket closure, and more importantly because of how readiness is substantially different that something that has the ability to read from a pipe at its leisure.  Not sure if there's opportunities for better refactoring and code reuse there... it's quite possible that there's not, despite how tempting it might seem.