Basis of new OutputSystem stuff.  Discarding the obsession with the Translator interface; that really just wasn't working out well for writing processes (that Completor stuff was a mess).  Replacing it with a ChannelWriter interface.  While some OutputSystem implementations will of course use use bases other than nio/channel stuff, that's to be hidden at a different level than one where the Translator abstraction can be terribly helpful.
Subtracted code from the commented-out stuff in FlowAssembler that the new parts of this commit evolved from.  Eclipse version Indigo-SR1 does crash on trying to syntax highlight source with generic types with multiple bounds, incidentally, which is a bit of a bummer.  SR2 still fails to highlight and dumps exceptions to stdout, but at least the whole program doesn't hang.