adding SLF4J.
I'll be using this mostly in the Test module.  Not sure if it'll show up in any of the other modules any time soon; for the most part I tend to see AHSlib as provided such low level components that logging is more relevant in the application above than in AHSlib itself.  Then again, it would seem like a reasonable thing for some parts of the IO system to be able to expose, or say TranslatingWorkTarget for example.