Create new WorkTarget to do closing of a channel.
It's a pretty short body, to the point that one might wonder if it's really deserving of its own entire WorkTarget.  It is.  Here's why: it's capable of failing.  So, if you tried to roll it into the same WorkTarget as handles the rest of the channel's actions, you start to have issues if for example the an exception is thrown from a channel read and then the close attempt also throws an exception: where do you put them both?  It's also something where essentially all of the logic about when it should occur is already expressed by using a completion listener, but those are forbidden to throw exceptions.  So, expressing this little action as its own WorkTarget dodges all of these punches nicely: it has a reasonable place to put exceptions, it reuses existing event patterns instead of requiring exotic triggering actions (which would also be tightly binding to a ChannelReader implementation!), and you can even wait on it if you like.